grpc_service/requirements.txt
grpcio
grpcio-tools
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
fastapi
uvicorn
